4.8.18 dev v4 (#28)
* Added percentage details to event batch attributes

* Added percentage details to event batch attributes

* Added config changes to process event type in cert pre processor

* 4.8.18 dev v4 (#24)

* added code for event cert generate

* added code for event cert generate

* added code for event cert generate

* added code for event cert generate

* added code for event cert generate

* added code for event cert generate

* added code for event cert generate

* worked on code review comments

* worked on code review comments

* worked on code review comments

* worked on code review comments

---------


* Fix for type cast error

* Using event batch instead of course_batch table for events

* Added fix for event cert notification

---------

+++ credential-generator/event-cert-generator/src/main/scala/org/event/job/certgen/domain/Event.scala
@@ -62,6 +62,6 @@ class Event(eventMap: java.util.Map[String, Any], partition: Int, offset: Long) 
 
   def eventId: String = related.getOrElse("eventId", "").asInstanceOf[String]
 
-  def eventCompletionPercentage: Int = readOrDefault[Int]("edata.eventCompletionPercentage", 0)
+  def eventCompletionPercentage: Double = readOrDefault[Double]("edata.eventCompletionPercentage", 0)
 }
